API: Remove 'immer' dependency
'immer' didn't really work for us and led to all kinds of hard-to-debug
problems. Instead of relying on 'immer' to handle changes during event
sourcing, we now carefully deep-copy relevant data in the
`*_eventsourcing.ts` files and have the event-specific modules `mutate`
the aggregate in-place. This makes the latter very simple; at the same
time we ensure that the aggregates stored in the cache aren't changed
outside the cache.

Affected aggregates:

- project
- subproject
- workflowitem