ui: Create new common component for additional data
The additional data component was specified only for subprojects. Since
projects and workflowitems can also contain additional data and the view
is the same for all of them, it made sense to create a common component
and move the logic for displaying additional data there.
The old component was removed from the SubProjectContainer and from the
codebase.