ui: Add additional data button for projects and workflowitems
ui: Add additional data button for projects

In the project overview, a button for displaying additional data is
displayed if the additional data is not empty.

ui: Add additional data button for workflowitems

Add button for workflowitems that displays additional data. The icon is
only displayed if there is additional data to display. This is to
prevent too many icons to be displayed.

ui: Show additional data button only if data is available

To be consistent, the additional data button is only displayed if there
is additional data to display on subproject level.

On subproject level, the additional data button was disabled when the
additional data was empty. Since the button is not displayed anymore if
the additional data is empty, this line was redundant.

Update CHANGELOG to add entry for issue #91

Resolve conflict for changelog