Fix and stabilize e2e-tests & provisioning
Some e2e-tests were failing, due to the following reasons:
* Issue with dayJS, which was fixed by a version update
* Missing user names - fixed by adding a fallback value
* Click does not open assignee selection (project and subproject) - fixed by identifying correct
element and waiting for it to be enabled
* Click does not open assignee selection (workflow item) - fixed by not
rendering batch edit drawer if no items are selected
* Missing projectId in API call for history drawer - fixed by using
already available projectId property

Since the process of updating a workflow item was changed in the API,
the provisioning ran into an error while performing an update of a
workflowitem.
The provisioning process was changed: The amount type was changed to
"N/A" while an amount is present, which only throws an error if the
workflow item is being closed. The workflow item is then reset to its
original state, so the provisioning error does not occur anymore.

Co-authored-by: Kevin Bader <kevin.bader@accenture.com>
Co-authored-by: Mathias Höld <mathias.hoeld@accenture.com>