ui: Fetch history items correctly
Change implementation of history fetching:

- The new approach features the concept of "pages" for the infinite
scroll elementi,i.e. each call fetches the subsequent page of the history
- The variables "offset" and "limit" are removed from the state, since
they can be calculated before each call based on the page that is
fetched
- The "initialLoad" for the inifinite scroll component has been
activated, so the component takes care of the initial fetch
- The "hasMore" variable is removed because the check if the last page
is reached can be done by each component itself
- The variable "historyItemsCount" has been renamed to
"totalHistoryItemCount"

Update formatting of history items:

* When a "projected_budget_updated" or "projected_budget_deleted" event
was generated for a project/subproject/workflowitem, the event was not
displayed correctly in the history list. This was fixed by adding an
entry in the stringifyHistory method and corresponding entries in the
language files.
* When a "grant_permission" or "revoke_permission" event was generated,
the granted/revoked permission was not displayed. This was fixed by
properly formatting the permission so it can be looked up in the
language file.

Fixes:

- Rename 'showHistory' to 'openHistory' to be consistent
- Remove dependency on ImmutableJS objects in HistoryList component and
pass plain JS objects to the component
- Simplify calculation of new offset
- Fix calculation of limit for the case where the totalHistoryItemCount
is a multiple of the historyPageSize
- Remove unnecessary elements from the history related state selectors
- Fix calculation for the "hasMore" property to prevent overflow
- Created new variable to check if the history call is the first one, so
the code is clearer
- HistoryList: Check if the business event and the snapshot exist to
prevent the UI from crashing
- Sagas: Change the calculation of the limit to increase readability

Add changelog entry