ui: Don't display amount if no exchange rate is set
When a workflow item is updated via the API from amount type "N/A" to
"disbursed" or "allocated", the exchange rate does not need to be set
immediately (it is only needed if the workflowitem is closed). Therefore
it can happen that no exchange rate (or amount is set).
If no amount or exchange rate is set, the amount field is left blank.