ui: Reset state for workflow item history when changing tabs
When a user views the workflow item details and clicks on history and
then back to another tab, the history state is now reset, so the
calculation of offset and limit are correct.

Add e2e test that makes sure this does not happen again.

Refactor e2e tests, so that 'data-test' is used instead of 'id'.

Closes #317

Rework based on review

- Moved WorkflowitemHistoryTab into Workflows
- Rewrote WorkflowitemHistory to class based component
- Removed effect hook for resetting history state and moved it to
ComponentWillUnmount
- Fixed e2e-tests
- Added LOGOUT action to WorkflowitemHistory reducer