ui: Only allow available budgets in subproject
When creating a subproject, the projected budget for a subproject can
only be selected from the projected budget of the parent project.
The organization is now a dropdown filled with the values of distinct
organizations in the project's projected budget.
Additionally, an error message is displayed if a currency is selected
that is already part of a projected budget for a given organization.
E2E tests are added that test this functionality.

Closes #228