ui: Add tags to projects
Tags have been added to projects to be able to provide additional data
to projects and prepare for a search functionality for projects.
Tags can be added when creating or editing a suproject. A new React
component "TagEditor" has been created to add and remove tags from a
project in the project dialog.

Tags are displayed in the project overview and on the project details
page, but only if there are any projects to display. The functionality
has been tested for backwards compatibility until version v1.0.0.

The documentation has been updated and the feature is described in more
detail in the 'Project' section of the user guide.

Closes #342