api: Add tags to projects
Tags are added to projects to be able to group projects by use case and
make metadata visible. This can also be used for a serch function later
on.
The tags have been added as new property for projects (instead of
additional data) to make handling of missing values easier and prevent
the additional data field from being misused.
Tags have been added to two endpoints and their corresponding workflows:
- global.createProject
- project.update
- project.list
- project.viewDetails

A regular expression has been added to projects to make sure that they
adhere to certain standards. The incoming tags are validated via JOI, such
that only unique values are saved and they are changed to lowercase.