api: grant/revoke global permissions only within organization
Global permissions should only be granted and revoked within the same
organization, i.e. a user from OrgaA should NOT be able to grant or
revoke permissios to/from users of OrgaB.
Additionally, the possibility to grant/revoke global permissions for
groups is removed, as this would cause inconsistencies and could be used
to grant/revoke global permissions across organizations.
Further changes:
 - Unit tests have been added to check for this behavior
 - Swagger documentation has been updated
 - The return type of the functions in the domain layer has been changed
to "Result"