ui: Add feature to filter projects in overview
A new feature is added which lets the user filter the projects in the
overview table. The search bar can be toggled by clicking the "Quick
Search" icon. The text entered into the search bar is used to filter the
projects that are visible to the user.
The "Cancel" button in the search bar can be used to clear the text. The
search bar is hidden (and reset) if a project details page is called.
The user can hit the "Escape" key to close and reset the search bar.

Add CHANGELOG entry and documentation for this new feature.

Closes #347