add stripUnknown options to Joi validation including unit tests
While parsing a streamitem into a business event on an old version,
it is important that do not ignore items with additional unknown properties.
Adding the "stripUnknown"-option to the validation which validates the first item of a "kind"
should prevent ignoring the whole resource.
First Eventtypes of the previous mentioned "kinds":
 - project_created
 - subproject_created
 - workflowitem_created
 - user_created
 - group_created
 - notification_created
 - global_permission_granted
 - global_permission_revoked
 - node_registered