api: Add restrictions for passwords
If a user wants to change his/her password, the new password has to
fulfill the following requirements:
* Minimum length of 8 characters
* At least one letter
* At least one number

The tests are added on Joi level and return a 'PreconditionError' if the
criteria are not met.

A unit test is added to test the conditions.

(cherry picked from commit 1aa0dc8d8e77d77c30f27c46a274eb3a9d76f23c)