e2e-tests: Fix a timing issue in permission tests
After executing the original Actions permissions may only be fetched after the changes are applied on the blockchain.
The changes are applied when the success action is dispatched which hides the loading indicator on the ui.
The loading indicator can be used to ensure that the changes are applied on the blockchain