API: Add global.disableUser and global.enableUser Intents
This endpoints will revoke or grant the user.authenticate permission.
If a user gets disabled, the user is not allowed to be assigned to
any project/subproject/workflowitem, otherwise an error-message
with information about these assignments is thrown. With disabling or
enabling, all permissions of the users remain unaffected except
'user.authenticate'.
Resolves #241