UI: Add feature to disable and enable users
Users can now be disabled and enabled at the /users site.
If an user gets disabled, this user dissappears everywhere except
the 'DISABLED USERS' tab. Before disabling an user, this user has
to be unassigned from all project/subproject/workflowitem.
If the user is still assigned to an
project/subproject/workflowitem, the snackbar shows up with an
error message.
Resolves #241